requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# wiki_extractor.py: 1
Wikipedia_API == 0.5.1

# content_processing.py: 2
nltk == 3.4

# content_processing.py: 7
pycorenlp == 0.3.0

# content_processing.py: 3
# main.py: 4
pyspotlight == 0.7.2

# content_processing.py: 5
# main.py: 8,9
rdflib == 4.2.2

# content_processing.py: 6
spacy == 2.0.18
